svc_test: Silence DecodeFrame() MSVC warnings.

+++ test/svc_test.cc
@@ -239,7 +239,7 @@ TEST_F(SvcTest, FirstFrameHasLayers) {
 
   const vpx_codec_err_t res_dec = decoder_->DecodeFrame(
       static_cast<const uint8_t *>(vpx_svc_get_buffer(&svc_)),
-      vpx_svc_get_frame_size(&svc_));
+      static_cast<int>(vpx_svc_get_frame_size(&svc_)));
 
   // this test fails with a decoder error
   ASSERT_EQ(VPX_CODEC_OK, res_dec) << decoder_->DecodeError();
@@ -268,7 +268,7 @@ TEST_F(SvcTest, EncodeThreeFrames) {
 
   vpx_codec_err_t res_dec = decoder_->DecodeFrame(
       static_cast<const uint8_t *>(vpx_svc_get_buffer(&svc_)),
-      vpx_svc_get_frame_size(&svc_));
+      static_cast<int>(vpx_svc_get_frame_size(&svc_)));
   ASSERT_EQ(VPX_CODEC_OK, res_dec) << decoder_->DecodeError();
 
   // FRAME 1
@@ -281,7 +281,7 @@ TEST_F(SvcTest, EncodeThreeFrames) {
 
   res_dec = decoder_->DecodeFrame(
       static_cast<const uint8_t *>(vpx_svc_get_buffer(&svc_)),
-      vpx_svc_get_frame_size(&svc_));
+      static_cast<int>(vpx_svc_get_frame_size(&svc_)));
   ASSERT_EQ(VPX_CODEC_OK, res_dec) << decoder_->DecodeError();
 
   // FRAME 2
@@ -294,7 +294,7 @@ TEST_F(SvcTest, EncodeThreeFrames) {
 
   res_dec = decoder_->DecodeFrame(
       static_cast<const uint8_t *>(vpx_svc_get_buffer(&svc_)),
-      vpx_svc_get_frame_size(&svc_));
+      static_cast<int>(vpx_svc_get_frame_size(&svc_)));
   ASSERT_EQ(VPX_CODEC_OK, res_dec) << decoder_->DecodeError();
 }
